requirements/_base.txt
django==1.6.5
django-grappelli==2.5.3
djangorestframework==2.3.14
south==1.0